relocated create and check caltable file
