+++ tests/test_casa.py
@@ -4,25 +4,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from flint.ms import MS
-from flint.selfcal.casa import args_to_casa_task_string, create_and_check_caltable_path
-
-
-def test_create_solution_path():
-    """Create the path to a solutions file, including appropriate handling of a channel range"""
-
-    ms_path = Path("/jack/has/a/measurement_set.ms")
-    cal_table = create_and_check_caltable_path(ms=MS.cast(ms_path))
-
-    assert isinstance(cal_table, Path)
-    assert cal_table == Path("/jack/has/a/measurement_set.caltable")
-
-    cal_table = create_and_check_caltable_path(
-        ms=MS.cast(ms_path), channel_range=(0, 143)
-    )
-
-    assert isinstance(cal_table, Path)
-    assert cal_table == Path("/jack/has/a/measurement_set.caltable.ch0000-0143")
+from flint.selfcal.casa import args_to_casa_task_string
 
 
 def test_args_to_casa_task_str():
